scheduler-scan/requirements.txt
redis==4.1.1
mysql-connector-python==8.0.28
python-dateutil==2.8.2